\name{f.analyzeIO.gui}
\alias{f.analyzeIO.gui}
\title{starts AnalyzeIO GUI}
\description{Starts an R/tk interfaced GUI that allows the user to
  explore an fMRI dataset stored in an ANALYZE format file using the
  function of the AnalyzeIO package.
}
\usage{
f.analyzeIO.gui()
}
\value{No value is returned
}
\keyword{utilities}
